--- web/html/src/manager/virtualization/pools/list/pools-list.renderer.js
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import { PoolsList } from './pools-list';
-import SpaRenderer from "core/spa/spa-renderer";
-
-export const renderer = (id, { serverId, pageSize, hypervisor }) => {
-  SpaRenderer.renderNavigationReact(
-    <PoolsList
-      pageSize={pageSize}
-      serverId={serverId}
-      hypervisor={hypervisor}
-    />,
-    document.getElementById(id),
-  );
-};
